feat: update other tables to use stored grid data
