+++ apps/trading/client-pages/portfolio/withdrawals-container.tsx
@@ -32,7 +32,7 @@ export const WithdrawalsContainer = () => {
         />
       </div>
       {!isReadOnly && (
-        <div className="h-auto flex justify-end px-[11px] py-2 bottom-0 right-3 absolute dark:bg-black/75 bg-white/75 rounded">
+        <div className="h-auto flex justify-end px-[11px] py-2 bottom-0 right-3 absolute rounded">
           <Button
             variant="primary"
             size="sm"
