+++ apps/trading/components/accounts-container/accounts-container.tsx
@@ -8,6 +8,10 @@ import { useVegaWallet } from '@vegaprotocol/wallet';
 import type { PinnedAsset } from '@vegaprotocol/accounts';
 import { AccountManager, useTransferDialog } from '@vegaprotocol/accounts';
 import { useDepositDialog } from '@vegaprotocol/deposits';
+import { create } from 'zustand';
+import { persist, subscribeWithSelector } from 'zustand/middleware';
+import type { ColumnState } from 'ag-grid-community';
+import { useDataGridStore } from '@vegaprotocol/datagrid';
 
 export const AccountsContainer = ({
   pinnedAsset,
@@ -23,6 +27,14 @@ export const AccountsContainer = ({
   const openWithdrawalDialog = useWithdrawalDialog((store) => store.open);
   const openDepositDialog = useDepositDialog((store) => store.open);
   const openTransferDialog = useTransferDialog((store) => store.open);
+
+  const [gridStore, update] = useAccountStore((store) => [
+    store.gridStore,
+    store.update,
+  ]);
+  const gridStoreCallbacks = useDataGridStore(gridStore, (colState) => {
+    update(colState);
+  });
 
   const onClickAsset = useCallback(
     (assetId?: string) => {
@@ -49,6 +61,7 @@ export const AccountsContainer = ({
         onMarketClick={onMarketClick}
         isReadOnly={isReadOnly}
         pinnedAsset={pinnedAsset}
+        gridProps={gridStoreCallbacks}
       />
       {!isReadOnly && !hideButtons && (
         <div className="flex gap-2 justify-end p-2 px-[11px] absolute lg:fixed bottom-0 right-3 dark:bg-black/75 bg-white/75 rounded">
@@ -72,3 +85,31 @@ export const AccountsContainer = ({
     </div>
   );
 };
+
+type Store = {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  filterModel?: { [key: string]: any };
+  columnState?: ColumnState[];
+};
+
+const useAccountStore = create<{
+  gridStore: Store;
+  update: (gridStore: Store) => void;
+}>()(
+  persist(
+    subscribeWithSelector((set) => ({
+      gridStore: {},
+      update: (newStore) => {
+        set((curr) => ({
+          gridStore: {
+            ...curr.gridStore,
+            ...newStore,
+          },
+        }));
+      },
+    })),
+    {
+      name: 'vega_accounts_store',
+    }
+  )
+);
