+++ apps/trading/components/fills-container/index.ts
@@ -1,0 +1,1 @@
+export * from './fills-container';
