+++ apps/trading/components/liquidity-container/index.ts
@@ -1,0 +1,1 @@
+export * from './liquidity-container';
