+++ apps/trading/components/positions-container/index.ts
@@ -1,0 +1,1 @@
+export * from './positions-container';
