+++ libs/accounts/src/lib/accounts-manager.tsx
@@ -11,10 +11,7 @@ import type { PinnedAsset } from './accounts-table';
 import { AccountTable } from './accounts-table';
 import { Dialog } from '@vegaprotocol/ui-toolkit';
 import BreakdownTable from './breakdown-table';
-import { create } from 'zustand';
-import { persist, subscribeWithSelector } from 'zustand/middleware';
-import type { ColumnState } from 'ag-grid-community';
-import { useDataGridStore } from '@vegaprotocol/datagrid';
+import type { useDataGridStore } from '@vegaprotocol/datagrid';
 
 const AccountBreakdown = ({
   assetId,
@@ -106,6 +103,7 @@ interface AccountManagerProps {
   onMarketClick?: (marketId: string, metaKey?: boolean) => void;
   isReadOnly: boolean;
   pinnedAsset?: PinnedAsset;
+  gridProps: ReturnType<typeof useDataGridStore>;
 }
 
 export const AccountManager = ({
@@ -116,17 +114,10 @@ export const AccountManager = ({
   isReadOnly,
   pinnedAsset,
   onMarketClick,
+  gridProps,
 }: AccountManagerProps) => {
   const gridRef = useRef<AgGridReact | null>(null);
   const [breakdownAssetId, setBreakdownAssetId] = useState<string>();
-  const [gridStore, update] = useAccountStore((store) => [
-    store.gridStore,
-    store.update,
-  ]);
-  const gridStoreCallbacks = useDataGridStore(gridStore, (colState) => {
-    update(colState);
-  });
-
   const { data, error } = useDataProvider({
     dataProvider: aggregatedAccountsDataProvider,
     variables: { partyId },
@@ -154,7 +145,7 @@ export const AccountManager = ({
         isReadOnly={isReadOnly}
         pinnedAsset={pinnedAsset}
         overlayNoRowsTemplate={error ? error.message : t('No accounts')}
-        {...gridStoreCallbacks}
+        {...gridProps}
       />
       <AccountBreakdownDialog
         assetId={breakdownAssetId}
@@ -167,29 +158,3 @@ export const AccountManager = ({
 };
 
 export default memo(AccountManager);
-
-type Store = {
-  filterModel?: { [key: string]: any };
-  columnState?: ColumnState[];
-};
-const useAccountStore = create<{
-  gridStore: Store;
-  update: (gridStore: Store) => void;
-}>()(
-  persist(
-    subscribeWithSelector((set) => ({
-      gridStore: {},
-      update: (newStore) => {
-        set((curr) => ({
-          gridStore: {
-            ...curr.gridStore,
-            ...newStore,
-          },
-        }));
-      },
-    })),
-    {
-      name: 'vega_accounts_store',
-    }
-  )
-);
