+++ libs/accounts/src/lib/accounts-table.tsx
@@ -306,7 +306,6 @@ export const AccountTable = forwardRef<AgGridReact, AccountTableProps>(
       <AgGrid
         {...props}
         style={{ width: '100%', height: '100%' }}
-        overlayNoRowsTemplate={t('No accounts')}
         getRowId={({ data }: { data: AccountFields }) => data.asset.id}
         ref={ref}
         tooltipShowDelay={500}
