+++ libs/fills/src/index.ts
@@ -1,3 +1,3 @@
-export * from './lib/fills-container';
+export * from './lib/fills-manager';
 export * from './lib/fills-data-provider';
 export * from './lib/__generated__/Fills';
