+++ libs/fills/src/lib/fills-manager.tsx
@@ -2,7 +2,7 @@ import type { AgGridReact } from 'ag-grid-react';
 import { useRef } from 'react';
 import { t } from '@vegaprotocol/i18n';
 import { FillsTable } from './fills-table';
-import { useBottomPlaceholder } from '@vegaprotocol/datagrid';
+import type { useDataGridStore } from '@vegaprotocol/datagrid';
 import { useDataProvider } from '@vegaprotocol/data-provider';
 import type * as Schema from '@vegaprotocol/types';
 import { fillsWithMarketProvider } from './fills-data-provider';
@@ -11,12 +11,14 @@ interface FillsManagerProps {
   partyId: string;
   marketId?: string;
   onMarketClick?: (marketId: string, metaKey?: boolean) => void;
+  gridProps: ReturnType<typeof useDataGridStore>;
 }
 
 export const FillsManager = ({
   partyId,
   marketId,
   onMarketClick,
+  gridProps,
 }: FillsManagerProps) => {
   const gridRef = useRef<AgGridReact | null>(null);
   const filter: Schema.TradesFilter | Schema.TradesSubscriptionFilter = {
@@ -36,9 +38,6 @@ export const FillsManager = ({
     },
     variables: { filter },
   });
-  const bottomPlaceholderProps = useBottomPlaceholder({
-    gridRef,
-  });
 
   return (
     <FillsTable
@@ -46,8 +45,8 @@ export const FillsManager = ({
       rowData={data}
       partyId={partyId}
       onMarketClick={onMarketClick}
-      {...bottomPlaceholderProps}
       overlayNoRowsTemplate={error ? error.message : t('No fills')}
+      {...gridProps}
     />
   );
 };
