+++ libs/ledger/src/index.ts
@@ -1,3 +1,2 @@
-export * from './lib/ledger-container';
 export * from './lib/ledger-manager';
 export * from './lib/__generated__/LedgerEntries';
