+++ libs/ledger/src/lib/ledger-table.tsx
@@ -61,8 +61,6 @@ export const LedgerTable = forwardRef<AgGridReact, LedgerEntryProps>(
             buttons: ['reset'],
           },
         }}
-        suppressLoadingOverlay
-        suppressNoRowsOverlay
         {...props}
       >
         <AgGridColumn
