+++ libs/orders/src/lib/components/index.ts
@@ -1,5 +1,4 @@
 export * from './order-data-provider';
 export * from './order-list';
 export * from './order-list-manager';
-export * from './order-list-container';
 export * from './mocks/generate-orders';
