+++ libs/positions/src/index.ts
@@ -1,6 +1,6 @@
 export * from './lib/__generated__/Positions';
-export * from './lib/positions-container';
 export * from './lib/positions-data-providers';
 export * from './lib/positions-table';
+export * from './libs/positions-manager';
 export * from './lib/use-market-margin';
 export * from './lib/use-open-volume';
